Added main page for renter.
