+++ app/src/main/java/com/example/barivara/renter/MainPage.java
@@ -2,49 +2,21 @@ package com.example.barivara.renter;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.widget.TextView;
+import android.view.View;
 
 import androidx.appcompat.app.AppCompatActivity;
 
-import com.example.barivara.MainActivity;
 import com.example.barivara.R;
-import com.google.android.gms.common.api.Status;
-import com.google.android.gms.location.places.Place;
-import com.google.android.gms.location.places.ui.PlaceAutocompleteFragment;
-import com.google.android.gms.location.places.ui.PlaceSelectionListener;
 
 public class MainPage extends AppCompatActivity {
 	@Override
 	protected void onCreate(Bundle savedInstanceState) {
 		super.onCreate(savedInstanceState);
 		setContentView(R.layout.activity_renter_main_page);
-
-		final TextView txtVw = findViewById(R.id.placeName);
-
-		PlaceAutocompleteFragment autocompleteFragment = (PlaceAutocompleteFragment)
-				getFragmentManager().findFragmentById(R.id.place_autocomplete_fragment);
-
-        /*AutocompleteFilter filter = new AutocompleteFilter.Builder()
-                .setCountry("IN")
-                .setTypeFilter(AutocompleteFilter.TYPE_FILTER_CITIES)
-                .build();
-        autocompleteFragment.setFilter(filter);*/
-		autocompleteFragment.setOnPlaceSelectedListener(new PlaceSelectionListener() {
-			@Override
-			public void onPlaceSelected(Place place) {
-				txtVw.setText(place.getName());
-			}
-			@Override
-			public void onError(Status status) {
-				txtVw.setText(status.toString());
-			}
-		});
 	}
 
-	@Override
-	public void onBackPressed() {
-		super.onBackPressed();
-		Intent intent = new Intent(this, MainActivity.class);
+	public void findAHouse(View view) {
+		Intent intent = new Intent(this,FindAHouse.class);
 		startActivity(intent);
 	}
 }
