+++ app/src/main/res/layout/activity_landlord_main_page.xml
@@ -34,7 +34,7 @@
 			android:id="@+id/textView8"
 			android:layout_width="wrap_content"
 			android:layout_height="wrap_content"
-			android:text="@string/amarBashagulo"
+			android:text="@string/amarBashagulo_Bariwala"
 			android:textAppearance="@style/TextAppearance.AppCompat.Large" />
 	</LinearLayout>
