+++ app/src/main/res/layout/activity_renter_find_a_house.xml
@@ -1,0 +1,36 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+	xmlns:app="http://schemas.android.com/apk/res-auto"
+	xmlns:tools="http://schemas.android.com/tools"
+	android:layout_width="match_parent"
+	android:layout_height="match_parent"
+	android:background="@color/defaultBackground"
+	tools:context=".renter.FindAHouse" >
+
+	<fragment
+		android:id="@+id/place_autocomplete_fragment"
+		android:name="com.google.android.gms.location.places.ui.PlaceAutocompleteFragment"
+		android:layout_width="match_parent"
+		android:layout_height="wrap_content"
+		android:layout_marginStart="16dp"
+		android:layout_marginLeft="16dp"
+		android:layout_marginTop="84dp"
+		android:layout_marginEnd="16dp"
+		android:layout_marginRight="16dp"
+		app:layout_constraintEnd_toEndOf="parent"
+		app:layout_constraintHorizontal_bias="0.0"
+		app:layout_constraintStart_toStartOf="parent"
+		app:layout_constraintTop_toTopOf="parent"
+		tools:layout="@layout/place_autocomplete_fragment" />
+
+	<TextView
+		android:id="@+id/placeName"
+		android:layout_width="wrap_content"
+		android:layout_height="wrap_content"
+		android:text="@string/houseInfo"
+		android:textSize="20sp"
+		app:layout_constraintBottom_toBottomOf="parent"
+		app:layout_constraintLeft_toLeftOf="parent"
+		app:layout_constraintRight_toRightOf="parent"
+		app:layout_constraintTop_toTopOf="parent" />
+</androidx.constraintlayout.widget.ConstraintLayout>
