+++ app/src/main/res/layout/activity_renter_main_page.xml
@@ -5,32 +5,67 @@
 	android:layout_width="match_parent"
 	android:layout_height="match_parent"
 	android:background="@color/defaultBackground"
-	tools:context=".renter.MainPage" >
+	tools:context=".renter.MainPage">
 
-	<fragment
-		android:id="@+id/place_autocomplete_fragment"
-		android:name="com.google.android.gms.location.places.ui.PlaceAutocompleteFragment"
-		android:layout_width="match_parent"
-		android:layout_height="wrap_content"
-		android:layout_marginStart="16dp"
-		android:layout_marginLeft="16dp"
-		android:layout_marginTop="84dp"
-		android:layout_marginEnd="16dp"
-		android:layout_marginRight="16dp"
-		app:layout_constraintEnd_toEndOf="parent"
-		app:layout_constraintHorizontal_bias="0.0"
-		app:layout_constraintStart_toStartOf="parent"
-		app:layout_constraintTop_toTopOf="parent"
-		tools:layout="@layout/place_autocomplete_fragment" />
-
-	<TextView
-		android:id="@+id/placeName"
+	<LinearLayout
 		android:layout_width="wrap_content"
 		android:layout_height="wrap_content"
-		android:text="@string/houseInfo"
-		android:textSize="20sp"
+		android:layout_marginStart="8dp"
+		android:layout_marginLeft="8dp"
+		android:layout_marginTop="80dp"
+		android:layout_marginEnd="8dp"
+		android:layout_marginRight="8dp"
+		android:layout_marginBottom="50dp"
+		android:clickable="true"
+		android:gravity="center"
+		android:onClick="findAHouse"
+		android:orientation="vertical"
+		app:layout_constraintBottom_toTopOf="@+id/linearLayout"
+		app:layout_constraintEnd_toEndOf="parent"
+		app:layout_constraintStart_toStartOf="parent"
+		app:layout_constraintTop_toTopOf="parent">
+
+		<ImageView
+			android:id="@+id/imageView3"
+			android:layout_width="200dp"
+			android:layout_height="200dp"
+			app:srcCompat="@drawable/ic_my_houses_2" />
+
+		<TextView
+			android:id="@+id/textView8"
+			android:layout_width="wrap_content"
+			android:layout_height="wrap_content"
+			android:text="@string/notunBashaVara"
+			android:textAppearance="@style/TextAppearance.AppCompat.Large" />
+	</LinearLayout>
+
+	<LinearLayout
+		android:id="@+id/linearLayout"
+		android:layout_width="wrap_content"
+		android:layout_height="wrap_content"
+		android:layout_marginStart="8dp"
+		android:layout_marginLeft="8dp"
+		android:layout_marginEnd="8dp"
+		android:layout_marginRight="8dp"
+		android:layout_marginBottom="80dp"
+		android:clickable="true"
+		android:gravity="center"
+		android:orientation="vertical"
 		app:layout_constraintBottom_toBottomOf="parent"
-		app:layout_constraintLeft_toLeftOf="parent"
-		app:layout_constraintRight_toRightOf="parent"
-		app:layout_constraintTop_toTopOf="parent" />
+		app:layout_constraintEnd_toEndOf="parent"
+		app:layout_constraintStart_toStartOf="parent">
+
+		<ImageView
+			android:id="@+id/imageView4"
+			android:layout_width="200dp"
+			android:layout_height="200dp"
+			app:srcCompat="@drawable/ic_my_houses_1" />
+
+		<TextView
+			android:id="@+id/textView9"
+			android:layout_width="wrap_content"
+			android:layout_height="wrap_content"
+			android:text="@string/amarBashagulo_Varatiya"
+			android:textAppearance="@style/TextAppearance.AppCompat.Large" />
+	</LinearLayout>
 </androidx.constraintlayout.widget.ConstraintLayout>
